# 리팩토링

See merge request s10-webmobile2-sub2/S10P12A109!41

Former-commit-id: e2ca2d47be4d710b0a39bf2d281f12b8bfeea417

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/feeds/request/FeedUpdateRequest.java
@@ -14,9 +14,9 @@ public class FeedUpdateRequest {
 
     private  String Authorizaiton;
     private String content;
-    private Long feedId;
-    private Long  clubId ;
-    private Long userId;
+    private int feedId;
+    private int  clubId ;
+    private int userId;
     private int visibility;
     private String imgName;
     private List<String> tags;
