+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/individualGoals/request/IndividualGoalUpdateRequest.java
@@ -11,8 +11,8 @@ import lombok.*;
 public class IndividualGoalUpdateRequest {
 
 
-    private Long individualGoalId; // 개인목표 기본키 번호
-    private Long userId;
+    private int individualGoalId; // 개인목표 기본키 번호
+    private int userId;
     private double kg;
     private double muscle;
     private double fat;
