+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/routine/response/CheckSuccessResponse.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @AllArgsConstructor
 @Builder
 public class CheckSuccessResponse {
-    private Long exerciseId;
+    private int exerciseId;
     private boolean success;
 }
