+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/repository/FavoriteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
     boolean existsByUserAndFeed(User user, Feed feed);
     Favorite findByUserAndFeed(User user, Feed feed);
